# backend/swagger.json
{
    "swagger": "2.0",
    "info": {
      "description": "Conduit API",
      "version": "1.0.0",
      "title": "Conduit API",
      "contact": {
        "name": "RealWorld",
        "url": "https://realworld.io"
      },
      "license": {
        "name": "MIT License",
        "url": "https://opensource.org/licenses/MIT"
      }
    },
    "host": "localhost:3000",
    "basePath": "/api",
    "schemes": [
      "http"
    ],
    "produces": [
      "application/json"
    ],
    "consumes": [
      "application/json"
    ],
    "securityDefinitions": {
      "Token": {
        "description": "For accessing the protected API resources, you must have received a a valid JWT token after registering or logging in. This JWT token must then be used for all protected resources by passing it in via the 'Authorization' header.\n\nA JWT token is generated by the API by either registering via /users or logging in via /users/login.\n\nThe following format must be in the 'Authorization' header :\n\n    Token: xxxxxx.yyyyyyy.zzzzzz\n    \n",
        "type": "apiKey",
        "name": "Authorization",
        "in": "header"
      }
    },
    "paths": {
...
      "/yuks": {
        "get": {
          "summary": "Get recent yuks globally",
          "description": "Get most recent yuks globally. Use query parameters to filter results. Auth is optional",
          "tags": [
            "yuks"
          ],
          "operationId": "Getyuks",
          "parameters": [
            {
              "name": "tag",
              "in": "query",
              "description": "Filter by tag",
              "required": false,
              "type": "string"
            },
            {
              "name": "author",
              "in": "query",
              "description": "Filter by author (username)",
              "required": false,
              "type": "string"
            },
            {
              "name": "favorited",
              "in": "query",
              "description": "Filter by favorites of a user (username)",
              "required": false,
              "type": "string"
            },
            {
              "name": "limit",
              "in": "query",
              "description": "Limit number of yuks returned (default is 20)",
              "required": false,
              "default": 20,
              "type": "integer"
            },
            {
              "name": "offset",
              "in": "query",
              "description": "Offset/skip number of yuks (default is 0)",
              "required": false,
              "default": 0,
              "type": "integer"
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/MultipleyuksResponse"
              }
            },
            "401": {
              "description": "Unauthorized"
            },
            "422": {
              "description": "Unexpected error",
              "schema": {
                "$ref": "#/definitions/GenericErrorModel"
              }
            }
          }
        },
        "post": {
          "summary": "Create an yuk",
          "description": "Create an yuk. Auth is required",
          "tags": [
            "yuks"
          ],
          "security": [
            {
              "Token": []
            }
          ],
          "operationId": "Createyuk",
          "parameters": [
            {
              "name": "yuk",
              "in": "body",
              "required": true,
              "description": "yuk to create",
              "schema": {
                "$ref": "#/definitions/NewyukRequest"
              }
            }
          ],
          "responses": {
            "201": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/SingleyukResponse"
              }
            },
            "401": {
              "description": "Unauthorized"
            },
            "422": {
              "description": "Unexpected error",
              "schema": {
                "$ref": "#/definitions/GenericErrorModel"
              }
            }
          }
        }
      },
...
      "/yuks/{slug}": {
        "get": {
          "summary": "Get an yuk",
          "description": "Get an yuk. Auth not required",
          "tags": [
            "yuks"
          ],
          "operationId": "Getyuk",
          "parameters": [
            {
              "name": "slug",
              "in": "path",
              "required": true,
              "description": "Slug of the yuk to get",
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/SingleyukResponse"
              }
            },
            "422": {
              "description": "Unexpected error",
              "schema": {
                "$ref": "#/definitions/GenericErrorModel"
              }
            }
          }
        },
        "put": {
          "summary": "Update an yuk",
          "description": "Update an yuk. Auth is required",
          "tags": [
            "yuks"
          ],
          "security": [
            {
              "Token": []
            }
          ],
          "operationId": "Updateyuk",
          "parameters": [
            {
              "name": "slug",
              "in": "path",
              "required": true,
              "description": "Slug of the yuk to update",
              "type": "string"
            },
            {
              "name": "yuk",
              "in": "body",
              "required": true,
              "description": "yuk to update",
              "schema": {
                "$ref": "#/definitions/UpdateyukRequest"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/SingleyukResponse"
              }
            },
            "401": {
              "description": "Unauthorized"
            },
            "422": {
              "description": "Unexpected error",
              "schema": {
                "$ref": "#/definitions/GenericErrorModel"
              }
            }
          }
        },
        "delete": {
          "summary": "Delete an yuk",
          "description": "Delete an yuk. Auth is required",
          "tags": [
            "yuks"
          ],
          "security": [
            {
              "Token": []
            }
          ],
          "operationId": "Deleteyuk",
          "parameters": [
            {
              "name": "slug",
              "in": "path",
              "required": true,
              "description": "Slug of the yuk to delete",
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            },
            "401": {
              "description": "Unauthorized"
            },
            "422": {
              "description": "Unexpected error",
              "schema": {
                "$ref": "#/definitions/GenericErrorModel"
              }
            }
          }
        }
      },
...
    },
    "definitions": {
...
      "Profile": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "bio": {
            "type": "string"
          },
          "image": {
            "type": "string"
          },
          "following": {
            "type": "boolean"
          }
        },
        "required": [
          "username",
          "bio",
          "image",
          "following"
        ]
      },
      "Yuk": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "image": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "liked": {
            "type": "boolean"
          },
          "disliked": {
            "type": "boolean"
          },
          "likesCount": {
            "type": "integer"
          },
          "dislikesCount": {
            "type": "integer"
          },
          "author": {
            "$ref": "#/definitions/Profile"
          }
        },
        "required": [
          "slug",
          "content",
          "image",
          "createdAt",
          "updatedAt",
          "liked",
          "disliked",
          "likesCount",
          "dislikesCount",
          "author"
        ]
      },
...
      "SingleyukResponse": {
        "type": "object",
        "properties": {
          "Yuk": {
            "$ref": "#/definitions/Yuk"
          }
        },
        "required": [
          "yuk"
        ]
      },
...
      "MultipleyuksResponse": {
        "type": "object",
        "properties": {
          "Yuks": {
            "type": "array",
            "items": {
              "$ref": "#/definitions/Yuk"
            }
          },
          "yuksCount": {
            "type": "integer"
          }
        },
        "required": [
          "yuks",
          "yuksCount"
        ]
      },
...
      "GenericErrorModel": {
        "type": "object",
        "properties": {
          "errors": {
            "type": "object",
            "properties": {
              "body": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            },
            "required": [
              "body"
            ]
          }
        },
        "required": [
          "errors"
        ]
      }
    }
  }